Project tests were corrected and renamed
